fix: dashboard dispatch context clone and serializer file move
Three small follow-ups from #5934.

DashboardDispatchContext.clone() raised a TypeError because its
own_properties left out the request its __init__ requires. Its
validate_filter_search_sort_fields also took no arguments while the base
declares (fields, refinement), which the two guarded call sites in
local_baserow/mixins.py pass.

service_serializers.py is merged into serializers.py, since we keep
serializers in one file per module.
